Rename `VolatilePtr` to `VolatilePtrSend`
