+++ src/lib.rs
@@ -14,11 +14,11 @@
 #![deny(unsafe_op_in_unsafe_fn)]
 
 pub use cell::VolatileCell;
-pub use ptr::VolatilePtr;
 pub use ptr_copy::VolatilePtrCopy;
+pub use ptr_send::VolatilePtr;
 
 pub mod access;
 mod cell;
 mod macros;
-mod ptr;
 mod ptr_copy;
+mod ptr_send;
